# backend/gen/go/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/proto/sbom/v1/component.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/sbom/decompose": {
      "post": {
        "operationId": "SbomService_Decompose",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SBOMDecomposition"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1DecomposeOptions"
            }
          }
        ],
        "tags": [
          "SbomService"
        ]
      }
    },
    "/api/v1/sbom/{fileUuid}/decompose": {
      "post": {
        "operationId": "SbomService_Decompose2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SBOMDecomposition"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "fileUuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SbomServiceDecomposeBody"
            }
          }
        ],
        "tags": [
          "SbomService"
        ]
      }
    },
    "/api/v1/sbom/{projectUid}/upload": {
      "post": {
        "operationId": "SbomService_Upload",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Revision"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "projectUid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SbomServiceUploadBody"
            }
          }
        ],
        "tags": [
          "SbomService"
        ]
      }
    }
  },
  "definitions": {
...
    "SbomServiceDecomposeBody": {
      "type": "object",
      "properties": {
        "upload": {
          "$ref": "#/definitions/v1SBOMFiles"
        },
        "maxDepth": {
          "type": "string",
          "format": "uint64"
        },
        "onlyVulnerable": {
          "type": "boolean"
        }
      }
    },
    "SbomServiceUploadBody": {
      "type": "object",
      "properties": {
        "upload": {
          "$ref": "#/definitions/v1SBOMFiles"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Advisory": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "v1Affect": {
      "type": "object",
      "properties": {
        "ref": {
          "type": "string"
        },
        "ranges": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Range"
          }
        }
      }
    },
    "v1Component": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "group": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "level": {
          "type": "integer",
          "format": "int32"
        },
        "bomRef": {
          "type": "string"
        },
        "purl": {
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Component"
          }
        },
        "vulnerabilities": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Vulnerability"
          }
        },
        "hasTransitiveVulns": {
          "type": "boolean"
        },
        "maxSeverity": {
          "type": "number",
          "format": "float"
        },
        "totalCveCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1Contact": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "bomRef": {
          "type": "string"
        }
      }
    },
    "v1DecomposeOptions": {
      "type": "object",
      "properties": {
        "fileUuid": {
          "type": "string"
        },
        "upload": {
          "$ref": "#/definitions/v1SBOMFiles"
        },
        "maxDepth": {
          "type": "string",
          "format": "uint64"
        },
        "onlyVulnerable": {
          "type": "boolean"
        }
      }
    },
    "v1DependencyCycle": {
      "type": "object",
      "properties": {
        "path": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1Lifecycle": {
      "type": "object",
      "properties": {
        "phase": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "v1Meta": {
      "type": "object",
      "properties": {
        "bomVersion": {
          "type": "string"
        },
        "tools": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Component"
          }
        },
        "project": {
          "$ref": "#/definitions/v1Component"
        },
        "authors": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Contact"
          }
        },
        "lifecycles": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Lifecycle"
          }
        },
        "properties": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "v1Range": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "range": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "v1Rating": {
      "type": "object",
      "properties": {
        "source": {
          "$ref": "#/definitions/v1Source"
        },
        "score": {
          "type": "number",
          "format": "float"
        },
        "severity": {
          "type": "string"
        },
        "method": {
          "type": "string"
        },
        "vector": {
          "type": "string"
        },
        "justification": {
          "type": "string"
        }
      }
    },
    "v1Revision": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "projectId": {
          "type": "string",
          "format": "uint64"
        },
        "gitTags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "gitBranch": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "gitCommitHash": {
          "type": "string"
        },
        "vcsUrl": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "sbomFileUuid": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1SBOMDecomposition": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "serialNumber": {
          "type": "string"
        },
        "md5": {
          "type": "string"
        },
        "metaData": {
          "$ref": "#/definitions/v1Meta"
        },
        "graph": {
          "$ref": "#/definitions/v1Component"
        },
        "components": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Component"
          }
        },
        "vulnerabilities": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Vulnerability"
          }
        },
        "totalNodes": {
          "type": "string",
          "format": "uint64"
        },
        "dependencyCycles": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1DependencyCycle"
          }
        }
      }
    },
    "v1SBOMFile": {
      "type": "object",
      "properties": {
        "fileName": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "v1SBOMFiles": {
      "type": "object",
      "properties": {
        "files": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1SBOMFile"
          }
        }
      }
    },
    "v1Source": {
      "type": "object",
      "properties": {
        "source": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "v1Vulnerability": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "source": {
          "$ref": "#/definitions/v1Source"
        },
        "description": {
          "type": "string"
        },
        "detail": {
          "type": "string"
        },
        "recommendation": {
          "type": "string"
        },
        "maxRating": {
          "type": "number",
          "format": "float"
        },
        "ratings": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Rating"
          }
        },
        "cwes": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "advisories": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Advisory"
          }
        },
        "affects": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Affect"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "publishedAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "rejectedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}